feat: Add webshop link functionality (#72)
feat: add webshop link functionality

This PR adds the functionality of making discounts clickable when there's a webshop property found in the discount. Configuration is updated with the ability to decide the target behavior of the `a` element of a discount, in which the link can be send the user directly on site (`_self`) or to a new tab (`_blank`). README is also updated with an example of how this functionality can be used.

Fixes https://github.com/Dintero/Dintero.Discounts.Web.SDK/issues/71